refactor master and user data model
