+++ Classes/Module/UserStorageModule.h
@@ -12,37 +12,9 @@
 #include <stdio.h>
 #include <dispatch/dispatch.h>
 #include <cocos2d.h>
-
 #include "json11.hpp"
 
-class UserItem
-{
-public:
-    static int index;
-    
-    int id, item_id, num;
-    UserItem() :
-        id(-1), item_id(0), num(0) {}
-    UserItem(int id, int item_id, int num) :
-        id(id), item_id(item_id), num(num) {}
-    UserItem(json11::Json::object data) :
-        id(data["id"].int_value()),
-        item_id(data["item_id"].int_value()),
-        num(data["num"].int_value()) {}
-    
-    json11::Json to_json() const
-    {
-        return json11::Json::object {
-            { "id", id },
-            { "item_id", item_id },
-            { "num", num }
-        };
-    }
-};
-
-const UserItem UserItemNull;
-
-class Item;
+class UserItem;
 class UserStorageModule
 {
 public:
